requirements.txt
fire==0.1.1
pyserial==3.4
